# docs/swagger.yaml
basePath: /
definitions:
...
  internal_handlers_contact.EmailInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_handlers_contact.SubscribeConfirmResponse:
    properties:
      list_name:
        type: string
      message:
        type: string
    type: object
  internal_handlers_contact.SuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
  internal_handlers_contact.UnsubscribeResponse:
    properties:
      list:
        type: string
      message:
        type: string
    type: object
...
  log_relay_internal_dtos.ForbiddenResponse:
    properties:
      error:
        example: You do not have permission to perform this action
        type: string
    type: object
  log_relay_internal_dtos.NotFoundErrorResponse:
    properties:
      error:
        example: Resource not found
        type: string
    type: object
  log_relay_internal_dtos.ServerErrorResponse:
    properties:
      error:
        example: An unexpected error occurred
        type: string
    type: object
  log_relay_internal_dtos.UnauthorizedResponse:
    properties:
      error:
        example: Invalid credentials or session
        type: string
    type: object
  log_relay_internal_dtos.ValidationErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        example: Validation failed
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Log Relay Project.
  title: Log Relay API
  version: "1.0"
paths:
...
  /api/subscriber/remove/{list_id}/confirm{token}:
    delete:
      consumes:
      - application/json
      description: Removes a contact from a mailing list using unsubscribe token
      parameters:
      - description: Public List ID
        in: path
        name: list_id
        required: true
        type: string
      - description: Unsubscribe Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers_contact.UnsubscribeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.UnauthorizedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.NotFoundErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.ServerErrorResponse'
      summary: Unsubscribe from mailing list
      tags:
      - subscribe
  /api/subscriber/signup/{list_id}/subscribe:
    post:
      consumes:
      - application/json
      description: Initiates email verification subscription flow for a mailing list
      parameters:
      - description: Public List ID
        in: path
        name: list_id
        required: true
        type: string
      - description: Email subscription input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_contact.EmailInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers_contact.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.ValidationErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.NotFoundErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.ServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.ServerErrorResponse'
      summary: Subscribe to mailing list
      tags:
      - subscribe
  /api/subscriber/signup/confirm:
    get:
      consumes:
      - application/json
      description: Confirms email subscription after verification
      parameters:
      - description: Public List ID
        in: query
        name: list_id
        required: true
        type: string
      - description: Verification Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers_contact.SubscribeConfirmResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.NotFoundErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.ServerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/log_relay_internal_dtos.ServerErrorResponse'
      summary: Confirm subscription
      tags:
      - subscribe
...
swagger: "2.0"